[CS] Allow much larger ids in postgres db
Some VIP users generate ids for their data that are larger than
Postgres's INTEGER type. This changes all id fields (and their
references) to be BIGINTs to accomodate that.

+++ resources/migrations/20150605-05-create-localities.up.sql
@@ -1,7 +1,7 @@
-CREATE TABLE localities (id INTEGER NOT NULL,
-                         results_id INTEGER REFERENCES results (id) NOT NULL,
+CREATE TABLE localities (id BIGINT NOT NULL,
+                         results_id BIGINT REFERENCES results (id) NOT NULL,
                          PRIMARY KEY (results_id, id),
                          name TEXT,
-                         state_id INTEGER,
+                         state_id BIGINT,
                          type TEXT,
-                         election_administration_id INTEGER);
+                         election_administration_id BIGINT);
