+++ resources/migrations/20150605-06-create-precincts.up.sql
@@ -1,9 +1,9 @@
-CREATE TABLE precincts(id INTEGER NOT NULL,
-                       results_id INTEGER REFERENCES results (id) NOT NULL,
+CREATE TABLE precincts(id BIGINT NOT NULL,
+                       results_id BIGINT REFERENCES results (id) NOT NULL,
                        PRIMARY KEY (results_id, id),
                        name TEXT,
                        number TEXT,
-                       locality_id INTEGER,
+                       locality_id BIGINT,
                        ward TEXT,
                        mail_only BOOLEAN,
                        ballot_style_image_url TEXT);
