+++ resources/migrations/20150605-10-create-precinct-splits.up.sql
@@ -1,6 +1,6 @@
-CREATE TABLE precinct_splits(id INTEGER NOT NULL,
-                             results_id INTEGER REFERENCES results (id) NOT NULL,
+CREATE TABLE precinct_splits(id BIGINT NOT NULL,
+                             results_id BIGINT REFERENCES results (id) NOT NULL,
                              PRIMARY KEY (results_id, id),
                              name TEXT,
-                             precinct_id INTEGER,
+                             precinct_id BIGINT,
                              ballot_style_image_url TEXT);
