+++ resources/migrations/20150605-15-create-ballots.up.sql
@@ -1,7 +1,7 @@
-CREATE TABLE ballots (id INTEGER NOT NULL,
-                      results_id INTEGER REFERENCES results (id) NOT NULL,
+CREATE TABLE ballots (id BIGINT NOT NULL,
+                      results_id BIGINT REFERENCES results (id) NOT NULL,
                       PRIMARY KEY (results_id, id),
-                      referendum_id INTEGER,
-                      custom_ballot_id INTEGER,
+                      referendum_id BIGINT,
+                      custom_ballot_id BIGINT,
                       write_in BOOLEAN,
                       image_url TEXT);
