+++ resources/migrations/20150605-17-create-referendums.up.sql
@@ -1,5 +1,5 @@
-CREATE TABLE referendums (id INTEGER NOT NULL,
-                          results_id INTEGER REFERENCES results (id) NOT NULL,
+CREATE TABLE referendums (id BIGINT NOT NULL,
+                          results_id BIGINT REFERENCES results (id) NOT NULL,
                           PRIMARY KEY (results_id, id),
                           title TEXT,
                           subtitle TEXT,
