+++ resources/migrations/20150605-21-create-state-early-vote-sites.up.sql
@@ -1,3 +1,3 @@
-CREATE TABLE state_early_vote_sites (results_id INTEGER REFERENCES results (id) NOT NULL,
-                                     state_id INTEGER,
-                                     early_vote_site_id INTEGER);
+CREATE TABLE state_early_vote_sites (results_id BIGINT REFERENCES results (id) NOT NULL,
+                                     state_id BIGINT,
+                                     early_vote_site_id BIGINT);
