+++ resources/migrations/20150605-25-create-locality-early-vote-sites.up.sql
@@ -1,3 +1,3 @@
-CREATE TABLE locality_early_vote_sites (results_id INTEGER REFERENCES results (id) NOT NULL,
-                                        locality_id INTEGER,
-                                        early_vote_site_id INTEGER);
+CREATE TABLE locality_early_vote_sites (results_id BIGINT REFERENCES results (id) NOT NULL,
+                                        locality_id BIGINT,
+                                        early_vote_site_id BIGINT);
