+++ resources/migrations/20150605-27-custom-ballots.up.sql
@@ -1,4 +1,4 @@
-CREATE TABLE custom_ballots (id INTEGER NOT NULL,
-                             results_id INTEGER REFERENCES results (id) NOT NULL,
+CREATE TABLE custom_ballots (id BIGINT NOT NULL,
+                             results_id BIGINT REFERENCES results (id) NOT NULL,
                              PRIMARY KEY (results_id, id),
                              heading TEXT);
